fix tests2
