+++ inspect_wandb/weave/autopatcher/__init__.py
@@ -1,0 +1,3 @@
+from inspect_wandb.weave.autopatcher.patcher import autopatch_inspect, reset_autopatch_inspect, get_inspect_patcher, CustomAutopatchSettings
+
+__all__ = ["autopatch_inspect", "reset_autopatch_inspect", "get_inspect_patcher", "CustomAutopatchSettings"]
